requirements.txt
python-telegram-bot>=21.0
pandas>=2.2.0
numpy>=1.26.0
ccxt>=4.2.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
sqlalchemy>=2.0.35
requests>=2.31.0
pytz>=2024.1
scipy>=1.11.0